Moved: Defaults class into the defaults package.
