+++ backend/src/main/java/br/balladesh/pjcappbackend/dto/api/ApiMessage.java
@@ -1,6 +1,6 @@
 package br.balladesh.pjcappbackend.dto.api;
 
-import br.balladesh.pjcappbackend.utilities.Defaults;
+import br.balladesh.pjcappbackend.utilities.defaults.Defaults;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.common.base.MoreObjects;
 import com.google.common.base.Objects;
