+++ backend/src/main/java/br/balladesh/pjcappbackend/dto/security/JwtJsonResponse.java
@@ -1,6 +1,6 @@
 package br.balladesh.pjcappbackend.dto.security;
 
-import br.balladesh.pjcappbackend.utilities.Defaults;
+import br.balladesh.pjcappbackend.utilities.defaults.Defaults;
 import com.google.common.base.MoreObjects;
 import com.google.common.base.Objects;
