+++ backend/src/test/java/br/balladesh/pjcappbackend/dto/api/artists/AddArtistRequestDTOTest.java
@@ -1,6 +1,6 @@
 package br.balladesh.pjcappbackend.dto.api.artists;
 
-import br.balladesh.pjcappbackend.utilities.Defaults;
+import br.balladesh.pjcappbackend.utilities.defaults.Defaults;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
