+++ backend/src/test/java/br/balladesh/pjcappbackend/dto/security/JwtJsonResponseTest.java
@@ -1,6 +1,6 @@
 package br.balladesh.pjcappbackend.dto.security;
 
-import br.balladesh.pjcappbackend.utilities.Defaults;
+import br.balladesh.pjcappbackend.utilities.defaults.Defaults;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
